Added demographic metric example for MPX project
This steps through creating a country and region based metric based on the demographic file and then merges with case data to produce a visualization (could also be added to a table).  You certainly don't have to do exactly this; it's only meant as an example which you can modify to meet your needs.